requirements.txt
fastapi
uvicorn
requests
python-dotenv
pydantic
beautifulsoup4
tldextract
groq